17.7.24

+++ mainwindow.cpp
@@ -3126,7 +3126,7 @@ void MainWindow::updateLabel()
 //           ui->label_vacactual->setText(QString::number(avg1));
         if(ui->label_vacpreset->text().toInt()>390)
         {
-         int dacval=ui->label_vacpreset->text().toInt()*16383/500;
+         int dacval=ui->label_vacpreset->text().toInt()*static_cast<int>(16383/500);
          l->writeDAC(dacval);
          qDebug()<<vac->convert(CHANNEL_1)<<dacval<<avgfp;
          int avg1 = vac->convert(CHANNEL_1)*0.1894;
